+++ simulator-core/src/main/java/com/evbox/everon/ocpp/simulator/configuration/SimulatorConfiguration.java
@@ -1,5 +1,6 @@
 package com.evbox.everon.ocpp.simulator.configuration;
 
+import com.evbox.everon.ocpp.simulator.station.StationHardwareData;
 import com.evbox.everon.ocpp.simulator.station.component.transactionctrlr.TxStartStopPointVariableValues;
 import com.evbox.everon.ocpp.v201.message.station.ConnectorStatus;
 import lombok.AllArgsConstructor;
@@ -40,6 +41,7 @@ public class SimulatorConfiguration {
         private Evse evse;
         private MeterValuesConfiguration meterValuesConfiguration = MeterValuesConfiguration.builder().build();
         private ComponentsConfiguration componentsConfiguration = ComponentsConfiguration.builder().build();
+        private HardwareConfiguration hardwareConfiguration = HardwareConfiguration.builder().build();
     }
 
     @Data
@@ -95,6 +97,12 @@ public class SimulatorConfiguration {
          */
         @Builder.Default
         private TransactionComponentConfiguration txCtrlr = TransactionComponentConfiguration.builder().build();
+
+        /**
+         * Authorisation controller componenet
+         */
+        @Builder.Default
+        private AuthComponentConfiguration authCtrl = AuthComponentConfiguration.builder().build();
     }
 
     @Data
@@ -106,6 +114,18 @@ public class SimulatorConfiguration {
          * Authentication password used for HTTP basic authentication
          */
         private String basicAuthPassword;
+    }
+
+    @Data
+    @Builder
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class AuthComponentConfiguration {
+        /**
+         * Station transaction authorisation toggle
+         */
+        @Builder.Default
+        private boolean authEnabled = true;
     }
 
     @Data
@@ -167,5 +187,23 @@ public class SimulatorConfiguration {
          */
         @Builder.Default
         private long pingIntervalMs = DEFAULT_PING_INTERVAL;
+    }
+
+    @Data
+    @Builder
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class HardwareConfiguration {
+        /**
+         * Station Vendor
+         */
+        @Builder.Default
+        private String vendor = StationHardwareData.VENDOR_NAME;
+
+        /**
+         * Station Model
+         */
+        @Builder.Default
+        private String model = StationHardwareData.MODEL;
     }
 }