+++ simulator-core/src/main/java/com/evbox/everon/ocpp/simulator/station/evse/states/RemotelyStoppedState.java
@@ -28,11 +28,11 @@ public class RemotelyStoppedState extends StoppedState {
         evse.stopTransaction();
         evse.clearToken();
 
-        stationMessageSender.sendStatusNotification(evse, evse.findConnector(connectorId));
         stationMessageSender.sendTransactionEventEnded(evseId, connectorId,
-                                                        TriggerReason.EV_DEPARTED,
-                                                        Reason.REMOTE,
-                                                        evse.getWattConsumedLastSession());
+                TriggerReason.EV_DEPARTED,
+                Reason.REMOTE,
+                evse.getWattConsumedLastSession());
+        stationMessageSender.sendStatusNotification(evse, evse.findConnector(connectorId));
 
         stateManager.setStateForEvse(evseId, new AvailableState());
         return CompletableFuture.completedFuture(UserMessageResult.SUCCESSFUL);